+++ src/main/java/com/doublew2w/sbs/mybatis/annotation/Update.java
@@ -1,0 +1,16 @@
+package com.doublew2w.sbs.mybatis.annotation;
+
+import java.lang.annotation.*;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/9/18 4:25
+ * @project: sbs-mybatis
+ */
+@Target(ElementType.METHOD)
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Update {
+  /** 查询SQL语句. */
+  String[] value();
+}