+++ src/main/java/com/doublew2w/sbs/mybatis/scripting/LanguageDriver.java
@@ -19,8 +19,15 @@ import org.dom4j.Element;
  * @project: sbs-mybatis
  */
 public interface LanguageDriver {
-  /** 创建SQL源码 */
+  /**
+   * 创建SQL源码
+   *
+   * <p>mapper xml方式
+   */
   SqlSource createSqlSource(Configuration configuration, Element script, Class<?> parameterType);
+
+  /** 创建SQL源码(annotation 注解方式) */
+  SqlSource createSqlSource(Configuration configuration, String script, Class<?> parameterType);
 
   /** 创建参数处理器 */
   ParameterHandler createParameterHandler(