+++ src/main/java/com/doublew2w/sbs/mybatis/scripting/xmltags/XMLLanguageDriver.java
@@ -6,6 +6,7 @@ import com.doublew2w.sbs.mybatis.mapping.MappedStatement;
 import com.doublew2w.sbs.mybatis.mapping.SqlSource;
 import com.doublew2w.sbs.mybatis.scripting.LanguageDriver;
 import com.doublew2w.sbs.mybatis.scripting.defaults.DefaultParameterHandler;
+import com.doublew2w.sbs.mybatis.scripting.defaults.RawSqlSource;
 import com.doublew2w.sbs.mybatis.session.Configuration;
 import org.dom4j.Element;
 
@@ -23,6 +24,14 @@ public class XMLLanguageDriver implements LanguageDriver {
     // 用XML脚本构建器解析
     XMLScriptBuilder builder = new XMLScriptBuilder(configuration, script, parameterType);
     return builder.parseScriptNode();
+  }
+
+  /** 用于处理注解配置 SQL 语句 */
+  @Override
+  public SqlSource createSqlSource(
+      Configuration configuration, String script, Class<?> parameterType) {
+    // 暂时不解析动态 SQL
+    return new RawSqlSource(configuration, script, parameterType);
   }
 
   @Override